restructure introducing framework
