--- integration/client.go
@@ -1,50 +1,0 @@
-// +build k8srequired
-
-package integration
-
-import (
-	"os"
-
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/credentials"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/ec2"
-	"github.com/giantswarm/microerror"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/tools/clientcmd"
-
-	"github.com/giantswarm/e2e-harness/pkg/harness"
-)
-
-func newK8sClient() (kubernetes.Interface, error) {
-	config, err := clientcmd.BuildConfigFromFlags("", harness.DefaultKubeConfig)
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-	cs, err := kubernetes.NewForConfig(config)
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-
-	return cs, nil
-}
-
-type aWSClient struct {
-	EC2 *ec2.EC2
-}
-
-func newAWSClient() aWSClient {
-	awsCfg := &aws.Config{
-		Credentials: credentials.NewStaticCredentials(
-			os.Getenv("AWS_ACCESS_KEY_ID"),
-			os.Getenv("AWS_SECRET_ACCESS_KEY"),
-			os.Getenv("AWS_SESSION_TOKEN")),
-		Region: aws.String(os.Getenv("AWS_REGION")),
-	}
-	s := session.New(awsCfg)
-	clients := aWSClient{
-		EC2: ec2.New(s),
-	}
-
-	return clients
-}
